Keep subGoals open on marking them complete
